Move template map into main, add integration test
* Move the template `FuncMap` into main.go
* Move the router mux into a function for testing
* Add a simple test against the default route
* Update Travis to test against multiple versions of Go

Closes #9

Note: Though not documented in the `README.md`, previously if a person
were to run the project without doing a build, they would see an error
like this:
```
go run main.go
./main.go:23: undefined: templateMap
```

With this commit you can run via `go run main.go` without doing a build
first. A person might wonder why `FuncMap` had to be moved into
'main.go` but not the contents of `rice-box.go`? The reason is that
`rice` is quite clever, and puts all of it's code in side `init()` ;)